google test file change

+++ test_googletest/src/sum_integers.hpp
@@ -1,3 +1,3 @@
 #pragma once
 #include <vector>
-int sum_integers(const std::vector<int> integers);
+double sum_integers(const std::vector<double> integers);
